Return ConduitError when Phabricator responds with HTML.
Sometimes Phabricator Conduit API can respond with HTML error instead of
JSON. For example when MySQL query is terminated due timeout - exception is
thrown. The exception is not handled by Conduit wrapper, errors are not
converted to JSON and HTML page is returned instead. This makes
difficult to investigate what was the actual response from Phab because
conversion from JSON error is returned. Another use case: if web server
(not Phabricator itself) responds with 429 (Too Many Requests) client is
not aware of status code.

This patch handles the situation. HTTP response code is provided as
`code` value of `ConduitError` struct and response body is provided in
`info` field.